fix tests, move Harmony tests from old to new
