+++ e2e/harmony/lane.e2e.1.ts
@@ -540,7 +540,7 @@ describe('bit lane command', function () {
       helper.fixtures.createComponentBarFoo();
       helper.fixtures.addComponentBarFooAsDir();
       helper.command.tagAllComponents();
-      helper.command.exportAllComponents();
+      helper.command.export();
       helper.scopeHelper.reInitLocalScopeHarmony();
       helper.scopeHelper.addRemoteScope();
       helper.command.createLane();
@@ -593,7 +593,7 @@ describe('bit lane command', function () {
       helper.command.tagAllComponents();
       helper.fixtures.createComponentBarFoo(fixtures.fooFixtureV2);
       helper.command.tagAllComponents();
-      helper.command.exportAllComponents();
+      helper.command.export();
 
       helper.scopeHelper.reInitLocalScopeHarmony();
       helper.bitJsonc.setupDefault();
@@ -627,7 +627,7 @@ describe('bit lane command', function () {
         helper.bitJsonc.setupDefault();
         helper.fixtures.populateComponents();
         helper.command.snapAllComponents();
-        helper.command.exportAllComponents();
+        helper.command.export();
 
         helper.command.createLane();
         helper.command.snapComponent('comp1 -f');
@@ -692,7 +692,7 @@ describe('bit lane command', function () {
         helper.command.createLane();
         helper.fixtures.populateComponents();
         helper.command.snapAllComponents();
-        helper.command.exportAllComponents();
+        helper.command.export();
       });
       it('as an intermediate step, make sure the lane is on the remote', () => {
         const lanes = helper.command.showRemoteLanesParsed();
@@ -737,7 +737,7 @@ describe('bit lane command', function () {
       helper.bitJsonc.setupDefault();
       helper.fixtures.populateComponents();
       helper.command.snapAllComponents();
-      helper.command.exportAllComponents();
+      helper.command.export();
 
       helper.command.createLane();
       helper.fixtures.populateComponents(undefined, undefined, ' v2');
@@ -800,7 +800,7 @@ describe('bit lane command', function () {
       helper.command.snapAllComponents();
     });
     it('should export with no errors about missing artifact files from the first snap', () => {
-      expect(() => helper.command.exportAllComponents()).to.not.throw();
+      expect(() => helper.command.export()).to.not.throw();
     });
   });
   describe('auto-snap when on a lane', () => {
@@ -853,7 +853,7 @@ describe('bit lane command', function () {
     // @todo
     describe.skip('importing the component to another scope', () => {
       before(() => {
-        helper.command.exportAllComponents();
+        helper.command.export();
 
         helper.scopeHelper.reInitLocalScopeHarmony();
         helper.scopeHelper.addRemoteScope();
